refactored away ModelAnalyzer, since it was basically just a constructor for ModelData
